registers and renders the background media block in php for fse
