+++ private/src/styles/blocks/text-media/_editor.scss
@@ -18,6 +18,7 @@
 .wp-block-amnesty-core-background-media > .block-editor-inner-blocks > .block-editor-block-list__layout {
   display: flex;
   width: 100%;
+  justify-content: space-around;
 
   .block-editor-block-list__block {
     flex-grow: 0;
