+++ wp-content/themes/humanity-theme/includes/blocks/register.php
@@ -12,6 +12,8 @@ require_once __DIR__ . '/action/register.php';
 require_once __DIR__ . '/action/render.php';
 require_once __DIR__ . '/background-media/register.php';
 require_once __DIR__ . '/background-media/render.php';
+require_once __DIR__ . '/background-media/components/register.php';
+require_once __DIR__ . '/background-media/components/render.php';
 require_once __DIR__ . '/blockquote/register.php';
 require_once __DIR__ . '/blockquote/render.php';
 require_once __DIR__ . '/collapsable/register.php';
@@ -68,6 +70,7 @@ if ( ! function_exists( 'amnesty_register_php_rendered_blocks' ) ) {
 	function amnesty_register_php_rendered_blocks() {
 		register_action_block();
 		register_background_media_block();
+		register_background_media_column_block();
 		register_blockquote_block();
 		register_collapsable_block();
 		register_custom_card_block();
